+++ crates/entrypoint/src/cli/command/azure/entra/user/mod.rs
@@ -1,12 +1,12 @@
 pub mod azure_entra_user;
-pub mod azure_entra_user_list;
 pub mod azure_entra_user_browse;
+pub mod azure_entra_user_list;
 
 pub use azure_entra_user::AzureEntraUserCommand;
-use clap::Args;
-use eyre::Result;
 pub use azure_entra_user_browse::AzureEntraUserBrowseArgs;
 pub use azure_entra_user_list::AzureEntraUserListArgs;
+use clap::Args;
+use eyre::Result;
 
 /// Entra user subcommands.
 #[derive(Args, Debug, Clone)]