+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Update `ct az devops audit` to include test plan license usage analysis
 - Update fetcher fn cache keys to fix uniqueness
 - Add `ct az ad user list` and `ct az ad user browse` commands
+- Add `ct az ad sp list` and `ct az ad sp browse` commands
 
 # v0.31.0